requirements.txt
albumentations==1.3.1
attrs==23.2.0
numpy==1.24.4
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.0.3
pytorch-lightning==2.1.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
torchmetrics==1.3.0.post0
